Improves network and I/O error handling
Enhances network communication robustness by using raw string literals for logging messages, improving readability and reducing potential escape sequence issues.

Corrects the data type for iteration cycle times to `u_long` to prevent potential overflow.

Improves I/O error handling and reporting in bus coupler communication, ensuring more reliable data acquisition and logging.

Returns an error code if any error occurs during reading input values.

+++ PAC/common/sys/dtime.h
@@ -65,8 +65,8 @@ struct stat_time
     u_long all_time;
     u_long cycles_cnt;
 
-    u_int  max_iteration_cycle_time;
-    u_int  min_iteration_cycle_time;
+    u_long  max_iteration_cycle_time;
+    u_long  min_iteration_cycle_time;
 
     int print_cycle_last_h;
